Create README.md
This WebApp provides insurance plan details of Multiple Users based on the search request. Given the Search request a Dynamic Query is generated which is used to retrieve the Data from The Data Base and provide the Search Result as Response in a Table. Additionally This project Also Implements Download Pdf And Download Excel File Functionality.